requirements.txt
### Python 3.11.5
absl-py==1.4.0
aiobotocore==2.5.0
aiohttp==3.8.4
aiohttp-retry==2.8.3
aioitertools==0.11.0
aiosignal==1.3.1
albumentations==1.3.1
amqp==5.1.1
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.2.1
async-lru==2.0.4
async-timeout==4.0.2
asyncssh==2.13.1
atpublic==3.1.2
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
backports.cached-property==1.0.2
beautifulsoup4==4.12.2
billiard==3.6.4.0
black==23.3.0
bleach==6.0.0
blessed==1.20.0
boto3==1.26.76
botocore==1.29.76
cachetools==5.3.1
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
clearml==1.11.0
clearml-agent==1.5.2
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cmake==3.26.3
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.3
configobj==5.0.8
contourpy==1.0.7
croniter==1.4.1
cryptography==41.0.1
cycler==0.11.0
dateutils==0.6.12
debugpy==1.6.7
decorator==5.1.1
deepdiff==6.6.0
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.6.1
distlib==0.3.6
distro==1.8.0
docker-pycreds==0.4.0
dpath==2.1.6
dulwich==0.21.5
dvc==2.58.2
dvc-data==0.51.0
dvc-http==2.30.2
dvc-objects==0.22.0
dvc-render==0.5.3
dvc-s3==2.22.0
dvc-studio-client==0.10.0
dvc-task==0.2.1
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.103.2
fastjsonschema==2.18.0
filelock==3.12.0
flatbuffers==23.5.26
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
funcy==2.0
furl==2.1.3
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.19.1
google-auth-oauthlib==1.0.0
gpustat==1.1
grandalf==0.8
grpcio==1.54.2
h11==0.14.0
huggingface-hub==0.17.3
humanfriendly==10.0
hydra-core==1.3.2
idna==3.4
imageio==2.30.0
imgaug==0.4.0
importlib-metadata==6.8.0
importlib-resources==6.0.0
iniconfig==2.0.0
inquirer==3.1.3
ipykernel==6.23.1
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
iterative-telemetry==0.0.8
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonschema==4.18.0
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.7.2
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
kiwisolver==1.4.4
kombu==5.2.4
lazy_loader==0.2
lightning==2.0.5
lightning-cloud==0.5.39
lightning-utilities==0.8.0
line-profiler==4.0.3
lit==16.0.5
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdurl==0.1.2
memory-profiler==0.61.0
mistune==3.0.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nanotime==0.5.2
nbclient==0.8.0
nbconvert==7.7.4
nbformat==5.9.2
nest-asyncio==1.5.6
networkx==3.1
notebook==7.0.2
notebook_shim==0.2.3
numpy #==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==11.525.112
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
nvitop==1.1.2
oauthlib==3.2.2
omegaconf==2.3.0
onnx #==1.14.0
onnxoptimizer==0.3.13
onnxruntime==1.15.0
opencv-python==4.8.0.74
opencv-python-headless==4.7.0.72
ordered-set==4.1.0
orderedmultidict==1.0.1
orjson==3.8.14
overrides==7.4.0
packaging==23.1
pandas==1.5.3
pandocfilters==1.5.0
panopticapi @ git+https://github.com/cocodataset/panopticapi.git@7bb4655548f98f3fedc07bf37e9040a992b054b0
parso==0.8.3
pathlib2==2.3.7.post1
pathspec==0.11.1
pathtools==0.1.2
pexpect==4.8.0
phenobench==0.1.0
pickleshare==0.7.5
Pillow==10.0.0
pkgutil_resolve_name==1.3.10
platformdirs==3.5.1
pluggy==1.0.0
prometheus-client==0.17.1
prompt-toolkit==3.0.38
protobuf==4.23.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
pycparser==2.21
pydantic==1.10.13
pydot==1.4.2
pygit2==1.12.1
Pygments==2.15.1
pygtrie==2.5.0
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.3.1
python-dateutil==2.8.2
python-editor==1.0.4
python-json-logger==2.0.7
python-multipart==0.0.6
pytorch-lightning==2.0.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
qudida==0.0.4
readchar==4.0.5
referencing==0.31.0
regex==2023.10.3
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.1
rpds-py==0.9.2
rsa==4.9
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
s3fs==2023.5.0
s3transfer==0.6.1
safetensors==0.4.0
scikit-image==0.20.0
scikit-learn==1.2.1
scipy==1.8.0
scmrepo==1.0.3
seaborn==0.12.2
Send2Trash==1.8.2
sentry-sdk==1.31.0
setproctitle==1.3.3
shapely==2.0.1
shortuuid==1.0.11
shtab==1.6.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
sqltrie==0.3.1
stack-data==0.6.2
starlette==0.27.0
starsessions==1.3.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.0
termcolor==2.3.0
termcolor-whl==1.1.3
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2023.4.12
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
tomlkit==0.11.8
torch==2.0.1
torch-ema==0.3
torch-tb-profiler==0.4.1
torchmetrics==0.10.3
torchsummary==1.5.1
torchvision==0.15.2
tornado==6.3.2
tplot==0.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.31.0
triton==2.0.0
types-python-dateutil==2.8.19.14
typing_extensions==4.6.2
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.23.2
vine==5.0.0
virtualenv==20.23.0
voluptuous==0.13.1
wandb==0.15.5
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.6.2
websockets==11.0.3
Werkzeug==2.3.4
widgetsnbextension==4.0.8
wrapt==1.15.0
yarl==1.9.2
zc.lockfile==3.0.post1
zipp==3.16.2
